use when instead of if, improve dep in pom, move BuildConstants to utils
