+++ MiraiMC-Sponge/src/main/templates/me/dreamvoid/miraimc/sponge/utils/BuildConstants.java
@@ -1,4 +1,4 @@
-package me.dreamvoid.miraimc.sponge;
+package me.dreamvoid.miraimc.sponge.utils;
 
 public class BuildConstants {
     public static final String VERSION = "${version}";
