+++ MiraiMC-Velocity/src/main/templates/me/dreamvoid/miraimc/velocity/utils/BuildConstants.java
@@ -1,4 +1,4 @@
-package me.dreamvoid.miraimc.velocity;
+package me.dreamvoid.miraimc.velocity.utils;
 
 public class BuildConstants {
     public static final String VERSION = "${version}";
